pdf_convert/requirements.txt
pdfplumber==0.10.3
pdf2image==1.16.3
pytesseract==0.3.10
Pillow==10.0.1
PyPDF2==3.0.1
poppler-utils==23.11.0

